python/requirements.txt
glfw~=2.9.0
numpy~=2.2.5
PyOpenGL~=3.1.9
pillow==11.2.1